Initial boiler plate
Introducing SynBlackList
------------------------

I was monitoring one of my server this morning and noticed many unsuccessful
ssh connection attempts in the log. Even when your passwords are strong I feel
this is still a nuisance. There are many solutions to this problem :

- either change the service port from the well known one to another one, which
  is not always feasible (some ports are well known for good reason).

- use a program that scans the log and automatically firewall connection
  attempts from this IP for a given probation period. It saves your server
some packets to receive and send and the bother for sshd to perform a full
blown handshake.

- any other solution that I'm unaware of.

This gave me the idea of a nice tutorial on how to write a plugin for Junkie.

As ideas usually come in pair, I also imagined to use git as a kind of generic
literate programming tool. Thus these unusually long commit messages.

So for a start I just created all the required boiler plate that you are
probably already too familiar with :

- autogen.sh : call all the autotools required to lift the versionned
  sourcetree into a development sourcetree

- configure.ac : preparing to use automake, autoheader, libtool and friends,
  and checking that the junkie header files are indeed present somewhere

- Makefile.am : main automakefile, just skipping to...

- src/Makefile.am : building the synbl.so plugin from...

- src/synbl.c, the plugin itself which performs nothing but to log a
  message in (de)initialization functions when called by junkie.

- .gitignore : informing git to ignore all generated files

Next commit we will sketch the internal of the plugin.