Initial design
So basically the plugin will measure the rate of TCP SYNs per client address.
When the rate raise above a given level, then a guile function is invoked
with the offending client address, with the idea that it's supposed to
configure the firewall to ban this IP altogether.  After a configurable
probation period another guile function is invoked for the same address, which
should then relieve the ban.

Why use a guile function for this ? Because it's the simplest way to make
configurable all host firewall software. For instance on Linux you will want
to run iptable while on openbsd you will want to run bpf, and on other host
you may want to do something completely different. This is nice to be able to
alter the behavior of this plugin without the need to recompile it.

So as we know we will need some scheme file, let's add a config directory.