The main data structures
This time we are merely defining two data structures.

The first one is a very basic struct syner which purpose is mainly to
store an IP address and a port, and which represent a remote client that
sent us a SYN to this port, SYNs that we are about to measure the rate.

This structure comes with for helping function to allocate, initialize,
deinitialize and free a struct syner.

As often when programming in C most of the changes are for laborious
repetitive coding. But at least it is simple and easy to follow.

The second structure, syners, is a hash of those syners, using syner.key
(ie. IP and port) as the indexing key. Notice that syner.key memset to 0
in syner constructor so that memcmp on hash keys will not be affected by
uninitialized padding bytes.

The hash is initialized (empty) when loading the plugin, and carefully
emptied and destructed at exit. This is because junkie being a dynamic
program, the plugin can be unloaded and the sniffer continuing its duty,
or even the same plugin reloaded several times in a loop, so you do not
want to leak memory at exit.

Next commit we will start performing actual work: receiving traffic
information and updating the hash of syners!