Accounting of syners
When the parse_callback() function is called (with the information on
the parser protocol stack and the packet itself), we first use the
ASSIGN_INFO family of MACROs to look the protocol stack for those we are
interested in:

- TCP because we are looking for TCP SYNs, and it gives us the
  destination port;

- IP because is gives us the sender address;

- CAP(ture) because it gives us the timestamp of packet reception;

Only if the packet matches these criteria will we pursue any further.

Then we merely build the proper syner_key (notice that we factorized a
syner_key_ctor so that we cannot forget the memset), formed by the
sender's IP and receiver port.

And then we lookup this key in the hash of already encountered syners.
If it's not there, we add it.

Regarding our syners hash, we will soon face two important questions :

- when do we forget a syner completely ?

- when do we resize the hash ?

We are going to address these questions in the two next commits.