Clear syners that does not exceed max SYN count
To measure the SYN rate we have a basic strategy:

- start to measure SYN count and create entries in the syners hash on
  reception of TCP SYN

- when a syners SYN count is greater than the allowed max, blacklist it

- after the sampling period is over, empty the hash and restart a new
  cycle.

So here we have added a second thread that merely sleep the duration of
the measurement, then awake and clears the hash. Later, we will
implement the SYN count.

Having another thread, we need a mutex to protect the access to the
syners hash (note that, despite having only local linkage, this lock is
called "synbl_lock" instead of merely "lock", for I like to have unique
names globally in order to reference easily any variable from a
debugger).

Notice also that the emptying of the hash was factored out into
syners_del_all() function.

At the start of the sleep/empty thread, which name is clearer_thread(),
notice the call to set_thread_name(). It's a good idea to name all
running threads so that you can use 'ps -eL' or even top (activating the
display of threads) to learn which thread consumes what portion of the
host resources. By convention in junkie all threads a prefixed with
"J-".

Next commit, we will add the SYN counters, so that we will be able to
sort the good syners from the ugly siners.