Calling extention function for blacklisting syners
The syn counting itself is straightforward : initialize a counter to 0
in the struct syner, then increment it each time a SYN is received.
Whenever the counter raises above the maximum allowed number of SYNs per
sampling period, blacklist this IP/port.

It was said previously that this blacklisting should be done from guile
so that it's easier to configure (even while the sniffer is running).
We won't see here how to call a guile function from C in great details,
so you'd better have a look to guile manual for that.

I will just draw your attention to some of the details.

First, notice that we return "(void *)1" from the C function called by
scm_with_guile(). This is to distinguish between a successful return
(when the instruction flow reach the "return (void *)1") from erroneous
returns caused by exceptions, which makes scm_with_guile returns NULL.

Second, notice that to lookup a guile function from it's name, you have
to lookup a variable in a module first, then dereference the variable to
have a pointer to the actual procedure required by scm_call_X().

Then, notice (in config/synbl.scm) that we defined our function in a
module (named "(junkie synbl)"). This is because you can't really know
what modules will be the current module when the function would be
defined, nor what module will be current when you call the
scm_c_lookup(). So it's easier to explicit all involved modules (or so I
think).

Now we have enough material to test the behavior of the plugin and the
next commit will be dedicated to automatic testing.