Addition of a simple functional test.
The first test of our test-suite merely plays a pcap file containing
many connection attempts and check that the offending client actually is
blacklisted (ie. that the synbl-ban function is called for this client).

The easiest way to test junkie is by writing the test in scheme and make
it evaluated by junkie itself, so that it's both easy to drive the test
and check the effect of the test on our plugin.

But to be able to do this we need to be able to setup the plugin
functions from guile and not only from the command line. So the proper
EXT_PARAM_RW are added to access the opt_* global variables. This made
me wonder if registering cli options should not automatically do this?

Anyway, let's go back to the test, in tests/test_bl.scm.
I could not come with tomething simpler that using sh as the shebang,
then running "junkie -c" from there. Ideally I'd like the shebang rules
to allow something like '#!/usr/bin/env junkie -c'.

So our ban unban function here merely update a global hashtable of
blacklisted clients/ports, so that at the end of the test, once the pcap
have been read, it's very easy to control this hash content.

There is only one feature left for synbl to be usable right now :
clients banishment must be relieved after the probation period. We will
add this in the next commit.