Remove clients from blacklist when probation period expires
The simplest way to call the unban function when the probation period
expires is to build a list of blacklisted clients (with the timestamp
when it was blacklisted), ordered in least recently blacklisted client,
and dedicate a thread to scan this list every seconds and free the
clients that were blacklisted for long enough.

Notice that because the list is ordered the scan can stop as soon as a
struct siner is met which is young enough to stay blacklisted.

Notice also how many lines of C are required for such a simple thing as
manipulate a mere list of elements that are dynamically allocated.

The test is then evolved to check that past the probation period the
blacklist is empty again.

As such synbl should be usable already, at least if we implement the
synbl-ban function to actually blacklist a client (and similarly for
synbl-unban). So next commit we will add the missing scheme bits.