First usable version --------------------
If by usable we mean that the program actually does something.

We only changed config/synbl.scm here, first to implement synbl-ban and
synbl-unban to use iptables to blacklist a given IP address. So if you run a
Linux OS this plugin should run and perform the required (un)banishments of
peers.

The scm file is also altered to load the plugin itself so that from junkie
command line you merely have to load this config file (assuming the library is
accessible to the loader).  So for instance, to run it from the build
directory, use:

LD_LIBRARY_PATH=src/.libs junkie -c config/synbl.scm

with additional options as suit your needs.

You should notice two important annoyances:

- we do not distinguish between incoming and outgoing traffic (and packets just
  passing by, actually), so that we may end up banishing our own host;

- many legitimate applications will soon get banished.

We will address these two issues in the two following commits.